pyproject.toml
[project]
name = "safety-knife"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pyspark (>=4.1.1,<5.0.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "yfinance (==1.2.0)",
    "delta-spark (==4.1.0)"
]

[tool.poetry]
packages = [{include = "safety_knife", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
